feat(DP_Alarm): add custom alarm music support

+++ Firmware/App/Service/DataProc/Def/Audio.h
@@ -30,16 +30,16 @@
 namespace DataProc {
 
 typedef struct Audio_Squence {
-    constexpr Audio_Squence(uint32_t freq = 0, uint32_t dur = 0, uint32_t time = 0)
+    constexpr Audio_Squence(uint16_t freq = 0, uint16_t dur = 0, uint32_t time = 0)
         : frequency(freq)
         , duration(dur)
         , time(time)
     {
     }
 
-    uint32_t frequency;
-    uint32_t duration;
-    uint32_t time;
+    uint16_t frequency;
+    uint16_t duration;
+    uint16_t time;
 } Audio_Squence_t;
 
 typedef struct Audio_Info {
